app/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
pymysql
python-multipart
python-jose[cryptography]
passlib[bcrypt]